fix simple build (depends on ocis-hello PR)
