+++ pkg/command/proxy.go
@@ -1,3 +1,5 @@
+// +build !simple
+
 package command
 
 import (
